+++ code/src/components/NavBar.vue
