+++ code/src/services/api.js
@@ -1,5 +1,7 @@
 import axios from 'axios';
 
-export const api = axios.create({
+const api = axios.create({
     baseURL: "http://localhost:3000/transport",
 })
+
+export default api;
